+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- Added a new environment variable `VITE_TRUST_URI_DOMAINS` to specify trusted domains for visualization requests
+
 ## [3.4.0](https://github.com/InseeFr/Drama-Queen/releases/tag/3.4.0) - 2026-05-20
 
 ### Changed
@@ -16,6 +20,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Authentication:
   - retrive scopes for jwt token
   - new env variable: `VITE_OIDC_SCOPES` (default value `"profile,roles"`)
+  - added authentication check on visualization page
 
 ## [3.3.2](https://github.com/InseeFr/Drama-Queen/releases/tag/3.3.2) - 2026-04-14